LayoutTests/platform/chromium-mac/fast/regex/early-acid3-86-expected.txt

Test that covers capturing brackets, and was adapted from a part of an early version of Acid3.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /TA[])]/.exec('TA]') threw exception SyntaxError: Invalid regular expression: /TA[])]/: Unmatched ')'.
PASS /[]/.exec('') is null
PASS /(\3)(\1)(a)/.exec('cat').toString() is 'a,,,a'
PASS successfullyParsed is true

TEST COMPLETE
